feat(tests): update Result<T> API usage patterns and activate tests
- Updated Result<T> API patterns in test files:
  - Changed result.error().code comparison to use static_cast<int>
  - Changed EXPECT_TRUE/FALSE(result) to use .is_ok()/.is_err()

- Activated 7 test files with 135 passing tests:
  - test_result_types.cpp
  - test_di_container.cpp
  - test_monitorable_interface.cpp
  - test_thread_context_simple.cpp
  - test_lock_free_collector.cpp
  - test_distributed_tracing.cpp
  - test_performance_monitoring.cpp

- Updated examples/basic_monitoring_example.cpp for Result<T> API

- Deferred 11 test files requiring missing headers:
  - optimization/, reliability/, utils/ headers not implemented
  - Documented in CMakeLists.txt comments

- Updated MON-002 ticket status to DONE in kanban

+++ tests/test_adaptive_monitoring.cpp
@@ -185,7 +185,7 @@ TEST_F(AdaptiveMonitoringTest, RegisterUnregisterCollector) {
     // Try to register again (should fail)
     result = monitor.register_collector("test", mock);
     ASSERT_FALSE(result.is_ok());
-    EXPECT_EQ(result.error().code, monitoring_error_code::already_exists);
+    EXPECT_EQ(result.error().code, static_cast<int>(monitoring_error_code::already_exists));
     
     // Unregister
     result = monitor.unregister_collector("test");
@@ -195,7 +195,7 @@ TEST_F(AdaptiveMonitoringTest, RegisterUnregisterCollector) {
     // Try to unregister non-existent (should fail)
     result = monitor.unregister_collector("test");
     ASSERT_FALSE(result.is_ok());
-    EXPECT_EQ(result.error().code, monitoring_error_code::not_found);
+    EXPECT_EQ(result.error().code, static_cast<int>(monitoring_error_code::not_found));
 }
 
 TEST_F(AdaptiveMonitoringTest, StartStopMonitoring) {
